ef poses through operator
