+++ bulletroboy/exoforce/exoforce_simulation.py
@@ -39,9 +39,6 @@ class ExoForceSim(ExoForce):
 		else:
 			raise Exception(f"Mode [{mode}] not supported!")
 
-		self.create_subscription(PoseStamped, Topics.OP_EF_POSES, self.operator_ef_pos_listener, 1)
-		
-	
 	def init_movement_params(self):
 		'''Initializes movement buttons of the GUI.
 		
@@ -243,29 +240,6 @@ class ExoForceSim(ExoForce):
 				tendon_sim = tendon
 				break
 		return tendon_sim
-
-	def operator_ef_pos_listener(self, ef_pose):
-		"""Callback of the pose subscriber. Sets the pose of the link given in the msg.
-
-		Args:
-			ef_pose: received PoseStamped msg.
-		
-		Returns:
-			-
-			
-		"""
-		ef_name = ef_pose.header.frame_id
-		#self.get_logger().info("Received pose for " + ef_name)
-		end_effector = self.get_ef_name(ef_name)
-		if end_effector is None:
-			self.get_logger().warn(ef_name + " is not an end effector!")
-			return
-
-		link_pos = np.array([ef_pose.pose.position.x, ef_pose.pose.position.y, ef_pose.pose.position.z])
-		link_orn = np.array([ef_pose.pose.orientation.x, ef_pose.pose.orientation.y, ef_pose.pose.orientation.z, ef_pose.pose.orientation.w])
-
-		end_effector.position = link_pos
-		end_effector.orientation = link_orn
 
 class TendonSim():
 	"""This class handles the simulation of each tendon attached to the operator.
