+++ bulletroboy/nodes/operator_node.py
@@ -1,12 +1,11 @@
 import rclpy
-from rclpy.executors import MultiThreadedExecutor
 from ..operator.operator_hardware import OperatorHW
 
 def main(args=None):
     rclpy.init(args=args)
 
     operator = OperatorHW()
-    rclpy.spin(operator, MultiThreadedExecutor())
+    rclpy.spin(operator)
 
     operator.destroy_node()
     rclpy.shutdown()
