+++ bulletroboy/operator/operator_hardware.py
@@ -1,11 +1,10 @@
-from rclpy.callback_groups import ReentrantCallbackGroup
+import time
 
 from geometry_msgs.msg import PoseStamped
 from roboy_simulation_msgs.msg import TendonUpdate
-from .operator import Operator, Link
-from ..utils.utils import Topics, Services
 
-import time
+from ..utils.utils import Services, Topics
+from .operator import Link, Operator
 
 class OperatorHW(Operator):
 	"""This class handles operator related functions and communicates directly with the TeleportVR app.
@@ -25,7 +24,7 @@ class OperatorHW(Operator):
 		self.pull_time = self.get_parameter("pull_time").get_parameter_value().double_value
 
 		self.target_force_publisher = self.create_publisher(TendonUpdate, Topics.TARGET_FORCE, 1)		
-		self.create_subscription(PoseStamped, Topics.VR_HEADSET_POSES, self.vr_pose_listener, 1, callback_group=ReentrantCallbackGroup())
+		self.create_subscription(PoseStamped, Topics.VR_HEADSET_POSES, self.vr_pose_listener, 1)
 
 		self.pull()
 		self.start_publishing()
